requirements.txt
https://www.djangoproject.com/download/1.7.b4/tarball/
django-allauth
django-model-utils
Markdown
